I AM ACTUALLY SO DUMB ! if i have moved my interfaces to DAL earlier.... I WOULD BE FINE !
