+++ DAL/DAL.csproj
@@ -15,8 +15,4 @@
     </PackageReference>
   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\BUS\BUS.csproj" />
-  </ItemGroup>
-
 </Project>
