+++ DAL/IRepos/Interface_TheThanhVien.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BUS.IServices
+namespace DAL.IRepos
 {
     public interface Interface_TheThanhVien
     {
